+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated to [GBDK](https://github.com/gbdk-2020/gbdk-2020) version 4.3.0
+- When using "Color Only" mode, GB rom file extension changed to ".gbc"
 
 ### Fixed